move sendVote, sendVideoIds, likeClicked, dislikeClicked to events.js
